+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.6.0 (in development)
+-----------------------
+- `get_local_repo()` now raises a dedicated `NoSuchRemoteError` if the given
+  remote does not exist
+
 v0.5.0 (2022-07-06)
 -------------------
 - Make `get_local_repo()` handle remote names that start with a hyphen